+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `<slot name="toolbar">` to `<leaflet-map>`
+- `<form is="share-target">` for handling shares using the Share Trarget API
 
 ## [v1.0.7] - 2020-07-20